+++ CHANGELOG.md
@@ -5,6 +5,11 @@ However, as this is still an experimental library, breaking changes may occur wi
 
 The official release will have a defined and more stable API. If you are already relying on a particular API, please let me know.
 
+## 0.3.3
+
+* Bug fixes
+  * Fix bug in retriggering that left next step in undefined state (#26)
+
 ## 0.3.2
 
 * Improvements: